Some small changes to stop warnings

+++ src/xlink2SystemELink.cpp
@@ -5,7 +5,7 @@ u64 SystemELink::getResourceVersion() const {
     return 0x1e;
 }
 
-u32 SystemELink::getUserParamNum() const {
+u64 SystemELink::getUserParamNum() const {
     return 0;
 }
